[FIX] *: Linter fixes
