+++ l10n_es_facturae/models/__init__.py
@@ -1,10 +1,9 @@
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl.html).
 
+from . import account_move
+from . import account_tax
+from . import account_tax_template
 from . import payment_mode
 from . import res_company
-from . import res_partner
-from . import account_tax_template
-from . import account_tax
 from . import res_currency
-from . import account_move
-from . import account_move_line
+from . import res_partner
